move helmholtz functions from recognition to new dreaming module
